Add files via upload
Project background and context

On its website, SpaceX lists Falcon 9 rocket launches at $62 million. Other providers charge upwards of $165 million for a single launch, but SpaceX's reusable first stage significantly reduces costs. Therefore, if an alternate company bids against SpaceX for a rocket launch, it needs to determine the cost of the launch. This project aims to create a machine-learning pipeline to predict the success of the first stage.

Problems you want to find answers

What factors determine if the rocket will land successfully?
The interaction among various features determines the success rate of a successful landing.
What operating conditions need to be in place to ensure a successful landing program?